run/backend/requirements.txt
Flask==3.0.0
gunicorn==20.1.0
google-cloud-firestore==2.16.0
pytz==2024.1
Flask-Cors==4.0.0